fix(network): exclude tunnels from broadcast set + fix startup race (R3)
Address two findings from `eval-network-adversary.md`:

Scenario E (red): `discovery/udp.rs::get_subnet_broadcasts` previously
enumerated all interfaces from `ifconfig` / `ip addr` output with NO
tunnel filter. Heartbeats then went out the VPN tunnel, so two laptops
on the same corporate VPN but different physical LANs would erroneously
discover each other.

  - Extracted `is_tunnel_interface` and `private_ipv4_rank` from
    `commands.rs` into a shared `net.rs` util module (used by both
    `commands.rs::best_effort_local_ip` and `discovery/udp.rs`).
  - Rewrote `get_subnet_broadcasts` on top of `if_addrs::get_if_addrs()`
    (already in Cargo.toml — no new dep) instead of shelling out to
    `ifconfig` / `ip addr`. Skips loopback and tunnel-named ifaces.
  - Extracted the broadcast-derivation step into a pure
    `compute_subnet_broadcasts` helper so it's unit-testable with
    synthetic iface lists; added 6 tests covering LAN, loopback,
    macOS utun, Windows AnyConnect, WireGuard / Tailscale, and the
    /0 limited-broadcast edge case.

Scenario D (yellow): `lib.rs::run` setup spawned the `disc_rx` consumer
thread AFTER `discovery_svc.start()`, leaving a race window where the
consumer's own `Database::open()` could lose to the setup-thread's
`mark_all_contacts_offline()`, making a freshly-upserted online row
get reset back to offline.

  - Reordered to: build channel → spawn consumer → start discovery,
    so the channel is being drained from t=0.

Verification:
- `cargo check` clean (15 pre-existing warnings, no new ones)
- `cargo test` 29/29 passed (was 23, +6 new broadcast tests; the
  Windows VPN / RFC1918 tests now live in `net::tests::` after the
  helper move)
- `pnpm run lint` clean (`tsc --noEmit` no errors)
- `pnpm run test` 43/43 passed (8 vitest files)
